task6/app/requirements.txt
fastapi
uvicorn
textblob
redis
psycopg2-binary
pydantic


